requirements.txt
Pillow==7.0.0
selenium==3.141.0
tqdm==4.43.0
urllib3==1.25.8
